refactor: relocate library primitives under ui and domain
